zig: Remove
The github release does not contain an asset with the binary. The
binaries are instead found on ziglang.org. This makes it hard to
automate keeping the package up to date, so it is removed for now.

Once a none github depended update routine has been implemented, this
package can be readded.